+++ .env
@@ -1,0 +1,1 @@
+EXTEND_ESLINT=true
